requirements.txt
image_classifiers==1.0.0
tensorflow-gpu==2.1.0
numpy==1.17.4
scipy==1.3.1
matplotlib==3.1.3
imgaug==0.4.0
seaborn==0.9.0
tqdm==4.40.0
git+git://github.com/tensorflow/examples.git@75d2c2a502f272977206b700f771dd6e7b901552
git+git://github.com/tensorpack/dataflow.git@9ffe010fb08ece4677c994247f1a5eeb2843f99e
git+git://github.com/qubvel/classification_models.git@e9364b85feacd38c511e41b1ae316c80595c8844
config_utils==0.3
tensorflow_datasets==1.3.0
Pillow==7.2.0
PyYAML==5.3.1
wandb==0.9.4
